tech-coe/learning/rama/python/atk-learning-rama-p1/pyproject.toml
[tool.poetry]
name = "atk-learning-rama-p1"
version = "0.1.0"
description = ""
authors = ["Ramarao Kanneganti <rama@aganitha.ai>"]
readme = "README.md"
packages = [{include = "atk_learning_rama_p1"}]

[tool.poetry.dependencies]
python = "^3.11"
simple-chalk = "^0.1.0"
typer = "^0.7.0"
requests = "^2.28.1"
pyyaml = "^6.0"

[tool.poetry.scripts]
atk-rama-p1-hello = "atk_learning_rama_p1.hello_world_commands:main"

[[tool.poetry.source]]
name = "prod"
url = "https://pypi.aganitha.ai/"
default = false
secondary = true


[[tool.poetry.source]]
name = "dev"
url = "https://dev-pypi.aganitha.ai/"
default = false
secondary = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
